+++ CHANGELOG.rst
@@ -34,6 +34,7 @@ Bugfixes
   which did not permanently remove the paths.
 - Fixed an issue in the ImageSeriesOperationsWindow where the correctness of 
   the output filename was not checked until after the operation.
+- Fixed an issue where Dataset axis ranges could be None.
 
 
 v24.03.25